Step 1: Extract helper functions from structures.py into parsing.py
Move _read_notes, _read_single_note, _parse_change_num, _process_note,
_process_settings, _process_normalize_tempo, _piece_process_normalize_tempo,
and copy_list into new musicpy/parsing.py module.

These functions are imported back into structures.py to maintain backward
compatibility. Uses lazy mp import in parsing.py to avoid circular dependency.
